+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 
 - **ext:** *(experimental)* An argument list with type assertions
+- **errors:** Add `Cause` method to contextual error
 
 
 ## 0.8.0 - 2017-08-06